+++ cmd/nvidia-ctk/runtime/configure/configure.go
@@ -20,13 +20,14 @@ import (
 	"fmt"
 	"path/filepath"
 
+	"github.com/urfave/cli/v2"
+
 	"github.com/NVIDIA/nvidia-container-toolkit/internal/logger"
 	"github.com/NVIDIA/nvidia-container-toolkit/pkg/config/engine"
 	"github.com/NVIDIA/nvidia-container-toolkit/pkg/config/engine/containerd"
 	"github.com/NVIDIA/nvidia-container-toolkit/pkg/config/engine/crio"
 	"github.com/NVIDIA/nvidia-container-toolkit/pkg/config/engine/docker"
 	"github.com/NVIDIA/nvidia-container-toolkit/pkg/config/ocihook"
-	"github.com/urfave/cli/v2"
 )
 
 const (
@@ -185,7 +186,7 @@ func (m command) validateFlags(c *cli.Context, config *config) error {
 		}
 	}
 
-	if config.runtime != "containerd" {
+	if config.runtime != "containerd" && config.runtime != "docker" {
 		if config.cdi.enabled {
 			m.logger.Warningf("Ignoring cdi.enabled flag for %v", config.runtime)
 		}
@@ -244,10 +245,9 @@ func (m command) configureConfigFile(c *cli.Context, config *config) error {
 		return fmt.Errorf("unable to update config: %v", err)
 	}
 
-	if config.cdi.enabled {
-		if err := cfg.Set("enable_cdi", true); err != nil {
-			return fmt.Errorf("failed enable CDI in containerd: %w", err)
-		}
+	err = enableCDI(config, cfg)
+	if err != nil {
+		return fmt.Errorf("failed to enable CDI in %s: %w", config.runtime, err)
 	}
 
 	outputPath := config.getOuputConfigPath()
@@ -299,4 +299,18 @@ func (m *command) configureOCIHook(c *cli.Context, config *config) error {
 		return fmt.Errorf("error creating OCI hook: %v", err)
 	}
 	return nil
+}
+
+// enableCDI enables the use of CDI in the corresponding container engine
+func enableCDI(config *config, cfg engine.Interface) error {
+	if !config.cdi.enabled {
+		return nil
+	}
+	switch config.runtime {
+	case "containerd":
+		return cfg.Set("enable_cdi", true)
+	case "docker":
+		return cfg.Set("experimental", true)
+	}
+	return fmt.Errorf("enabling CDI in %s is not supported", config.runtime)
 }